requirements.txt
argon2-cffi>=20.1.0
arrow>=1.1.0
async-generator>=1.10
attrs>=20.3.0
backcall>=0.2.0
bleach>=3.3.0
blis>=0.7.4
catalogue>=2.0.4
certifi>=2021.5.30
cffi>=1.14.5
chardet>=4.0.0
click>=7.1.2
-e git+https://github.com/CristianCantoro/compressed_stream.git@02a06eb1a2e6184de297c80552c1a8c741f99687#egg=compressed_stream
cymem>=2.0.5
decorator>=4.0.9
defusedxml>=0.7.1
docopt>=0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints>=0.3
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0-py3-none-any.whl
future>=0.18.2
fuzzywuzzy>=0.8.0
geocoder>=1.38.1
geographiclib>=1.50
geopy>=2.1.0
idna>=2.10
ipykernel>=5.5.3
ipython>=7.22.0
ipython-genutils>=0.2.0
ipywidgets>=7.6.3
it-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.0.0/it_core_news_sm-3.0.0-py3-none-any.whl
jedi>=0.18.0
Jinja2>=2.11.3
jsonable>=0.3.1
jsonschema>=3.2.0
jupyter>=1.0.0
jupyter-client>=6.1.12
jupyter-console>=6.4.0
jupyter-core>=4.7.1
jupyterlab-pygments>=0.1.2
jupyterlab-widgets>=1.0.0
liwc>=0.5.0
m3inference>=1.1.3
Mako>=1.0.2
MarkupSafe>=0.23
mathplot>=0.1
mediawiki-utilities>=0.4.18
mistune>=0.8.4
more-itertools>=6.0.0
murmurhash>=1.0.5
mwcites>=0.2.0
mwcli>=0.0.1
mwparserfromhell>=0.4.2
mwtypes>=0.2.0
mwxml>=0.2.0
nbclient>=0.5.3
nbconvert>=6.0.7
nbformat>=5.1.3
nest-asyncio>=1.5.1
networkx>=1.11
notebook>=6.3.0
numpy>=1.20.2
packaging>=20.9
pandas>=1.2.3
pandocfilters>=1.4.3
para>=0.0.5
parso>=0.8.2
pathy>=0.5.2
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=8.1.2
plotly>=4.14.3
preshed>=3.0.5
prometheus-client>=0.10.1
prompt-toolkit>=3.0.18
ptyprocess>=0.7.0
pycld2>=0.41
pycparser>=2.20
pydantic>=1.7.4
Pygments>=2.8.1
PyMySQL>=0.7.1
pyparsing>=2.4.7
pyrsistent>=0.17.3
python-dateutil>=2.8.1
python-Levenshtein>=0.12.0
pytz>=2021.1
pyzmq>=22.0.3
qtconsole>=5.0.3
QtPy>=1.9.0
ratelim>=0.1.6
rauth>=0.7.3
regex>=2018.8.17
requests>=2.25.1
retrying>=1.3.3
Send2Trash>=1.5.0
six>=1.15.0
smart-open>=3.0.0
spacy>=3.0.6
spacy-legacy>=3.0.6
srsly>=2.4.1
terminado>=0.9.4
testpath>=0.4.4
thinc>=8.0.5
torch>=1.8.1
torchvision>=0.9.1
tornado>=6.1
tqdm>=4.59.0
traitlets>=5.0.5
typer>=0.3.2
typing>=3.5.0.1
typing-extensions>=3.7.4.3
urllib3>=1.26.5
wasabi>=0.8.2
wcwidth>=0.2.5
webencodings>=0.5.1
widgetsnbextension>=3.5.1